% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addCylinder}
\alias{addCylinder}
\title{Create and add cylinder shape}
\usage{
addCylinder(id, spec = list())
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})}

\item{spec}{Style specification}
}
\value{
R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})
}
\description{
Create and add cylinder shape
}
