% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{addStyle}
\alias{addStyle}
\title{Add style properties to all selected atoms}
\usage{
addStyle(id, sel = list(), style = list())
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})}

\item{sel}{Atom selection specification}

\item{style}{style spec to add to specified atoms}
}
\value{
R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})
}
\description{
Add style properties to all selected atoms
}
