% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{bohr_cube}
\alias{bohr_cube}
\title{Gaussian cube file example}
\format{
Gaussian cube format
}
\source{
\url{https://github.com/3dmol/3Dmol.js/blob/master/tests/auto/data/bohr.cube}
}
\usage{
bohr_cube
}
\description{
Gaussian cube file example
}
\keyword{datasets}
