% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cif_254385}
\alias{cif_254385}
\title{Cif file example}
\format{
cif format
}
\source{
\url{https://github.com/3dmol/3Dmol.js/blob/master/tests/auto/data/254385.cif}
}
\usage{
cif_254385
}
\description{
Cif file example
}
\keyword{datasets}
