% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{cleanup_nulls}
\alias{cleanup_nulls}
\title{(Internal Use) Cleanup Nulls}
\usage{
cleanup_nulls(list)
}
\arguments{
\item{list}{A list() object to be cleaned.}
}
\description{
For the given list, if any of the defined items are NULL, removes it from
the list so it is not passed on to the callJS() function.
}
