% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{cube_benzene_homo}
\alias{cube_benzene_homo}
\title{Gaussian cube file example}
\format{
Gaussian cube format
}
\source{
\url{https://github.com/3dmol/3Dmol.js/blob/master/tests/test_structs/benzene-homo.cube}
}
\usage{
cube_benzene_homo
}
\description{
Gaussian cube file example
}
\keyword{datasets}
