% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{.m_add_line}
\alias{.m_add_line}
\title{Add Line Between Points}
\usage{
.m_add_line(id, start, end, dashed = FALSE, spec = m_shape_spec())
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from
\code{r3dmol()}).}

\item{start}{Start location of line Can be either \code{m_sel()} or
\code{m_vector3()}.}

\item{end}{End location of line. Can be either \code{m_sel()} or
\code{m_vector3()}.}

\item{dashed}{Boolean, whether or not to draw line as dashed.}

\item{spec}{Additional shape specifications defined with
\code{m_shape_spec()}.}
}
\value{
R3dmol \code{id} or a \code{r3dmol} object (the output from
\code{r3dmol()}).
}
\description{
Add a line between the given points. Used intrenally.
}
