% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multi.R
\name{.m_multi_spec}
\alias{.m_multi_spec}
\title{List of multiple start/stop locations.}
\usage{
.m_multi_spec(starts, ends, pairwise = FALSE)
}
\arguments{
\item{starts}{Start locations, either \code{m_vector3()} or \code{m_sel()}}

\item{ends}{End locations, either \code{m_vector3()} or \code{m_sel()}}

\item{pairwise}{Logical, TRUE returns all possible combinations of start
and end locations.}
}
\description{
For use in addition of lines, arrows, cylinders, when multiple
locations are specified in a single call.
}
