% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{isAnimated}
\alias{isAnimated}
\title{Get viewer animate status}
\usage{
isAnimated(id)
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})}
}
\value{
\code{logical}
}
\description{
Return true if viewer is currently being animated, false otherwise
}
