% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{m_add_as_one_molecule}
\alias{m_add_as_one_molecule}
\title{Create and add model to viewer}
\usage{
m_add_as_one_molecule(id, data, format)
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})}

\item{data}{Input data}

\item{format}{Input format}
}
\value{
R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})
}
\description{
Given multimodel file and its format, all atoms are added to one model
}
