% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{m_add_box}
\alias{m_add_box}
\title{Create and add box shape}
\usage{
m_add_box(id, spec = list())
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})}

\item{spec}{Box shape style specification}
}
\value{
R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})
}
\description{
This method provides a shorthand way to create a box shape object
}
\examples{
library(r3dmol)

r3dmol() \%>\%
  m_add_line(spec = list(
    color = 'red',
    start = m_vector3(0, 0, 0),
    end = m_vector3(5, 0, 0)
  )) \%>\%
  m_add_line(spec = list(
    color = 'blue',
    start = m_vector3(0, 0, 0),
    end = m_vector3(0, 5, 0)
  )) \%>\%
  m_add_line(spec = list(
    color = 'green',
    start = m_vector3(0, 0, 0),
    end = m_vector3(0, 0, 5)
  )) \%>\%
  m_add_box(spec = list(
    center = m_vector3(0, 0, 0),
    demensions = list(w = 3, h = 4, d = 2),
    color = "magenta"
  )) \%>\%
  m_zoom_to() \%>\%
  m_rotate(angle = 45, axis = m_vector3(1, 1, 1))
}
