% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{m_add_lines}
\alias{m_add_lines}
\title{Add Lines Between Points}
\usage{
m_add_lines(
  id,
  starts,
  ends,
  dashed = FALSE,
  pairwise = FALSE,
  spec = m_shape_spec()
)
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from
\code{r3dmol()}).}

\item{starts}{Single of list of starting positions. Can be either
\code{m_sel()} or \code{m_vector3()}.}

\item{ends}{Single of list of ending positions. Can be either
\code{m_sel()} or \code{m_vector3()}.}

\item{dashed}{Logical whether the lines are dashed.}

\item{pairwise}{Logical, TRUE creates lines between all possible combinations
of starts and ends points provided.}

\item{spec}{\code{m_shape_spec()} for line styling.}
}
\value{
R3dmol \code{id} or a \code{r3dmol} object (the output from
\code{r3dmol()})
}
\description{
Add lines between the given points.
}
\examples{
library(r3dmol)

r3dmol() \%>\%
  m_add_model(data = pdb_6zsl) \%>\%
  m_set_style(style = m_style_cartoon()) \%>\%
  m_zoom_to() \%>\%
  m_add_style(
    sel = m_sel(resi = 1:10),
    style = c(
      m_style_stick(),
      m_style_sphere(scale = 0.3)
    )
  ) \%>\%
  m_add_lines(
    starts = m_sel(resi = 1, chain = "A"),
    ends = list(m_sel(resi = 10, chain = "A"),
                m_sel(resi = 10, chain = "B")),
    dashed = TRUE
  )
}
