% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{m_add_model}
\alias{m_add_model}
\title{Create and add model to viewer}
\usage{
m_add_model(id, data, format = c("pdb", "sdf", "xyz", "pqr", "mol2"), options)
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})}

\item{data}{Path of input data path or a vector of data.}

\item{format}{Input format (\code{'pdb'}, \code{'sdf'}, \code{'xyz'}, \code{'pqr'},
or \code{'mol2'})}

\item{options}{Format dependent options. Attributes depend on the input file format.}
}
\value{
R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})
}
\description{
Create and add model to viewer, given molecular data and its format
}
\examples{
library(r3dmol)

r3dmol() \%>\%
  m_add_model(data = pdb_6zsl, format = "pdb")
}
