% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{m_add_outline}
\alias{m_add_outline}
\title{Add colored outline to all objects in scene.}
\usage{
m_add_outline(id, width = 0.1, color = "black")
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from
\code{r3dmol()})}

\item{width}{Width of the outline, defaults to 0.1}

\item{color}{Color of the outline, defaults to black.}
}
\description{
Adds a colored outline to all objects in the scene, helping the viewer
to distinguish depth in often complex molecular scenes.
}
\examples{
library(r3dmol)

r3dmol() \%>\%
  m_add_model(data = pdb_1j72) \%>\%
  m_set_style(style = m_style_stick()) \%>\%
  m_add_outline()
}
