% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{m_add_shape}
\alias{m_add_shape}
\title{Add shape object to viewer}
\usage{
m_add_shape(id, shapeSpec = list())
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})}

\item{shapeSpec}{Style specification for label}
}
\value{
R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})
}
\description{
Add shape object to viewer
}
