% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{m_add_style}
\alias{m_add_style}
\title{Add style properties to all selected atoms}
\usage{
m_add_style(id, sel = list(), style = list())
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})}

\item{sel}{Atom selection specification}

\item{style}{style spec to add to specified atoms}
}
\value{
R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})
}
\description{
Add style properties to all selected atoms
}
\examples{
library(r3dmol)

r3dmol() \%>\%
  m_add_model(data = pdb_1j72, format = "pdb") \%>\%
  m_add_style(style = list(cartoon = list())) \%>\%
  m_zoom_to()
}
