% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{m_add_volumetric_data}
\alias{m_add_volumetric_data}
\title{Construct isosurface from volumetric data in gaussian cube format}
\usage{
m_add_volumetric_data(id, data, format, or)
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})}

\item{data}{Input file contents}

\item{format}{Input file format}

\item{or}{Shape style specification}
}
\value{
R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})
}
\description{
Construct isosurface from volumetric data in gaussian cube format
}
\examples{
r3dmol() \%>\%
  m_add_volumetric_data(
    data = bohr_cube,
    format = "cube",
    or = list(
      isoval = -0.01,
      color = "red",
      opacity = 0.95
    )
  ) \%>\%
  m_set_style(sel = list(cartoon = list()),
              style = list(stick = list())) \%>\%
  m_zoom_to()
}
