% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bio3d.R
\name{m_bio3d}
\alias{m_bio3d}
\title{Load structure from package bio3d}
\usage{
m_bio3d(structure)
}
\arguments{
\item{structure}{bio3d object containing coordinates for desired structure}
}
\description{
Function to take bio3d structure and use in the r3dmol app.
}
