% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button.R
\name{m_button}
\alias{m_button}
\title{Add button into viewer}
\usage{
m_button(id, name, label, func, style = "flex-start")
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from
\code{r3dmol()}).}

\item{name}{Name for button.}

\item{label}{Label for button.}

\item{func}{The function executed when the button is clicked.}

\item{layout}{CSS style for button layout.}
}
\value{
R3dmol \code{id} or a \code{r3dmol} object (the output from
\code{r3dmol()})
}
\description{
Add button into viewer
}
\examples{
library(r3dmol)

r3dmol() \%>\%
  m_add_model(data = pdb_1j72, format = "pdb") \%>\%
  m_zoom_to() \%>\%
  m_button(
    name = "cartoon",
    label = "Cartoon",
    func = "
      function(){
        viewer.setStyle({cartoon:{}});
        viewer.render();
      }
    "
  )
}
