% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button.R
\name{m_button_add_surface}
\alias{m_button_add_surface}
\title{Add a button for adding a surface}
\usage{
m_button_add_surface(
  id,
  label = "Surface",
  type = "VDW",
  style = m_style_surface(),
  atomsel = m_sel(),
  allsel = NULL,
  removeSurface = TRUE,
  removeLabel = "Clear Surfaces"
)
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from
\code{r3dmol()}).}

\item{label}{String for button label.}

\item{type}{Surface type ('VDW', 'MS', 'SAS', or 'SES')}

\item{style}{Optional style specification for surface material created with \code{m_style_surface()} to alter color and opacity.}

\item{atomsel}{Show surface for atoms in this selection.}

\item{allsel}{Use different atoms to calculate the surface than to display
it. May be a larger selection than \code{atomsel}.}

\item{removeSurface}{Logical, whether to also create a button that will
remove all surfaces from the viewer.}

\item{removeLabel}{String for clear surface button label.}
}
\value{
R3dmol \code{id} or a \code{r3dmol} object (the output from
\code{r3dmol()})
}
\description{
Add a button for adding a surface
}
\examples{
# Add buttons to add and clear surface representations
r3dmol() \%>\%
  m_add_model(data = pdb_1j72, format = "pdb") \%>\%
  m_zoom_to() \%>\%
  m_set_style(m_style_cartoon()) \%>\%
  m_button_add_surface(
    style = m_style_surface(opacity = 0.4),
    atomsel = m_sel(resi = 70:110)
  ) \%>\%
  m_button_add_surface(
    style = m_style_surface(opacity = 1),
    atomsel = m_sel(resi = 70:110, invert = TRUE),
    label = "Rest of Protein",
    removeSurface = FALSE
  )
}
