% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button.R
\name{m_button_animate}
\alias{m_button_animate}
\title{Add a button for starting and stopping animations}
\usage{
m_button_animate(
  id,
  interval = 50,
  loop = "backAndForth",
  step = 1,
  reps = 0,
  label = "Play Animation",
  stopButton = TRUE,
  stopLabel = "Stop Animation"
)
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from
\code{r3dmol()}).}

\item{interval}{Time in milliseconds between frames.}

\item{loop}{Direction of animation loop. \code{'backward'}, \code{'forward'}
or \code{'backAndForth'}}

\item{step}{How many frames to step through. 1 plays every frame, 2 skips
every second frame etc.}

\item{reps}{How many times to repeat the animation. 0 loops indefinitely.}

\item{label}{String for button label.}

\item{stopButton}{Logical, whether to also create a button to stop the
currently playing animation.}

\item{stopLabel}{String for stop for stop animation button label.}
}
\value{

}
\description{
Add a button for starting and stopping animations
}
