% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button.R
\name{m_button_zoom_to}
\alias{m_button_zoom_to}
\title{Add a button for zooming to a selection}
\usage{
m_button_zoom_to(
  id,
  sel = m_sel(),
  label = "Zoom",
  duration = 500,
  zoomOut = TRUE,
  zoomOutLabel = "Zoom Out"
)
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from
\code{r3dmol()}).}

\item{sel}{Atom selection specification with \code{m_sel()}}

\item{label}{String for button label.}

\item{duration}{Duration of the zoom animation in milliseconds. (Default 500)}

\item{zoomOut}{Logical, whether to also create a button that will reset the
view to encompass the entire scene.}

\item{zoomOutLabel}{String for zoom out button label.}
}
\value{

}
\description{
Add a button for zooming to a selection
}
