% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{m_callback}
\alias{m_callback}
\title{Callback function generator}
\usage{
m_callback(elementId, callback)
}
\arguments{
\item{elementId}{The same \code{elementId} set in r3dmol() for bind the callback function.}

\item{callback}{Callback function in string.}
}
\description{
Callback function generator
}
\examples{
r3dmol(elementId = "pdb_6zsl") \%>\% # Required to provide elementId
  m_add_model(data = pdb_6zsl, format = "pdb") \%>\%
  m_set_style(style = list(cartoon = list(color = 'spectrum'))) \%>\%
  m_add_arrow(
    spec = list(
      start = list(x = -10.0, y = 0.0, z = 0.0),
      end = list(x = 0.0, y = -10.0, z = 0.0),
      radius = 1.0,
      radiusRadio = 1.0,
      mid = 1.0,
      clickable = TRUE,
      callback = m_callback(
        elementId = "pdb_6zsl", # The same elementId
        callback = "function(){
          this.color.setHex(0xFF0000FF);
        }"
      )
    )
  )
}
