% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{m_cartoon_style}
\alias{m_cartoon_style}
\title{Specify styling for Cartoon}
\usage{
m_cartoon_style(
  color = "default",
  style = "default",
  ribbon = F,
  arrows = F,
  tubes = F,
  thickness = 0.4,
  width = 1,
  opacity = 1
)
}
\arguments{
\item{color}{Strand color, may specify as 'spectrum' which will apply
reversed gradient based on residue number.}

\item{style}{style of cartoon rendering (trace, oval, rectangle (default),
parabola, edged)}

\item{ribbon}{whether to use constant strand width, disregarding secondary
structure; use thickness to adjust radius}

\item{arrows}{whether to add arrows showing beta-sheet directionality;
does not apply to trace or ribbon}

\item{tubes}{whether to display alpha helices as simple cylinders; does not
apply to trace}

\item{thickness}{cartoon strand thickness, default is 0.4}

\item{width}{cartoon strand width, default is secondary structure-dependent;
does not apply to trace or ribbon}

\item{opacity}{set opacity from 0-1; transparency is set per-chain with a
warning outputted in the event of ambiguity}
}
\description{
Returns a list for the \code{style = list(cartoon = list())} function,
for setting the cartoon specifications.
}
