% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{m_clear}
\alias{m_clear}
\title{Clear scene of all objects}
\usage{
m_clear(id)
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from
\code{r3dmol()})}
}
\value{
R3dmol \code{id} or a \code{r3dmol} object (the output from
\code{r3dmol()})
}
\description{
Clear scene of all objects
}
