% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{m_create_model_from}
\alias{m_create_model_from}
\title{Create a new model from atoms specified by sel}
\usage{
m_create_model_from(id, sel, extract)
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from
\code{r3dmol()})}

\item{sel}{Atom selection specification.}

\item{extract}{If true, remove selected atoms from existing models}
}
\value{
R3dmol \code{id} or a \code{r3dmol} object (the output from
\code{r3dmol()})
}
\description{
Create a new model from atoms specified by sel.
If extract, removes selected atoms from existing models.
}
