% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{m_get_model}
\alias{m_get_model}
\title{Return specified model}
\usage{
m_get_model(id, modelId)
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from
\code{r3dmol()})}

\item{modelId}{Retrieve model with specified id}
}
\value{
R3dmol \code{id} or a \code{r3dmol} object (the output from
\code{r3dmol()})
}
\description{
Return specified model
}
