% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get.R
\name{m_is_animated}
\alias{m_is_animated}
\title{Get viewer animate status}
\usage{
m_is_animated(id)
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from
\code{r3dmol()})}
}
\value{
\code{logical}
}
\description{
Return true if viewer is currently being animated, false otherwise
}
