% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{m_label_style}
\alias{m_label_style}
\title{Specify styling for Labels}
\usage{
m_label_style(
  font = NULL,
  fontSize = NULL,
  fontColor = NULL,
  fontOpacity = NULL,
  borderThickness = NULL,
  borderColor = NULL,
  borderOpacity = NULL,
  backgroundColor = NULL,
  backgroundOpacity = NULL,
  position = NULL,
  inFront = NULL,
  showBackground = NULL,
  fixed = NULL,
  alignment = c("topLeft", "topCenter", "topRight", "centerLeft", "center",
    "centerRight", "bottomLeft", "bottomCenter", "bottomRight"),
  frame = NULL
)
}
\arguments{
\item{font}{font name, default sans-serif}

\item{fontSize}{height of text, default 18}

\item{fontColor}{font color, default white}

\item{fontOpacity}{font opacity, default 1}

\item{borderThickness}{line width of border around label, default 0}

\item{borderColor}{color of border, default backgroundColor}

\item{borderOpacity}{opacity of border, default 1}

\item{backgroundColor}{color of background, default black}

\item{backgroundOpacity}{opacity of background, default 1}

\item{position}{x,y,z coordinates for label (for custom positioning)}

\item{inFront}{Logical, if \code{TRUE} always put in front of model}

\item{showBackground}{Logical, show background rounded rectangle, default \code{TRUE}}

\item{fixed}{Logical, setes the label to change with the model when zooming}

\item{alignment}{String, how to orient the label with respect to position: 'topLeft' (default), 'topCenter', 'topRight', 'centerLeft', 'center', 'centerRight', 'bottomLeft', 'bottomCenter', 'bottomRight'}
}
\description{
Returns a list for the \code{m_add_res_labels(style = list())} function,
for setting the label specifications.
}
