% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{m_line_style}
\alias{m_line_style}
\title{Specify styling for Lines}
\usage{
m_line_style(colorscheme = NULL, color = NULL, opacity = NULL, hidden = NULL)
}
\arguments{
\item{colorscheme}{element based coloring}

\item{color}{fixed coloring, overrides colorscheme}

\item{opacity}{opacity, must be the same for all atoms in the model}

\item{hidden}{Logical, do not show line}
}
\description{
Returns a list for the \code{style = list(line = list())} function
for setting the line specifications.
}
