% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/save.R
\name{m_png}
\alias{m_png}
\title{Convert widgets to PNG image}
\usage{
m_png(id, width, height)
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from
\code{r3dmol()}).}

\item{width, height}{image width and height.}
}
\value{
Base64 encoded png image wrapped by \code{<img>} tag.
}
\description{
Convert widgets to PNG image
}
\examples{
library(r3dmol)

r3dmol() \%>\%
  m_add_model(data = pdb_1j72, format = "pdb") \%>\%
  m_set_style(style = m_style_cartoon()) \%>\%
  m_zoom_to() \%>\%
  m_png(width = 600)
}
