% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{m_render}
\alias{m_render}
\title{Render current state of viewer}
\usage{
m_render(id)
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})}
}
\description{
Render current state of viewer, after adding/removing models,
applying styles, etc.
}
