% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{m_resize}
\alias{m_resize}
\title{Resize viewer according to containing HTML element's dimensions}
\usage{
m_resize(id)
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})}
}
\description{
Resize viewer according to containing HTML element's dimensions
}
\examples{
library(r3dmol)

mol <- r3dmol() \%>\%
  m_add_model(data = pdb_6zsl, format = "pdb") \%>\%
  m_zoom_to() \%>\%
  m_set_width(300)

mol

mol \%>\%
  m_resize()
}
