% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{m_set_color_by_element}
\alias{m_set_color_by_element}
\title{Set color by element}
\usage{
m_set_color_by_element(id, sel, colors)
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from
\code{r3dmol()})}

\item{sel}{Atom selection.}

\item{colors}{Color hex code or name.}
}
\value{
R3dmol \code{id} or a \code{r3dmol} object (the output from
\code{r3dmol()})
}
\description{
Set color by element
}
