% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{m_set_default_cartoon_quality}
\alias{m_set_default_cartoon_quality}
\title{Set the default cartoon quality for newly created models}
\usage{
m_set_default_cartoon_quality(id, quality)
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from
\code{r3dmol()})}

\item{quality}{Default cartoon quality.}
}
\value{
R3dmol \code{id} or a \code{r3dmol} object (the output from
\code{r3dmol()})
}
\description{
Set the default cartoon quality for newly created models. Default is
\code{5}.
Current models are not affected.
}
\examples{
library(r3dmol)

r3dmol() \%>\%
  m_set_default_cartoon_quality(20) \%>\%
  m_add_model(data = pdb_1j72, format = "pdb") \%>\%
  m_set_style(style = m_style_cartoon()) \%>\%
  m_zoom_to()
}
