% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{m_set_hover_duration}
\alias{m_set_hover_duration}
\title{Set the duration of the hover delay}
\usage{
m_set_hover_duration(id, hoverDuration)
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from
\code{r3dmol()})}

\item{hoverDuration}{an optional parameter that denotes
the duration of the hover delay (in milliseconds) before the hover action is
called}
}
\value{
R3dmol \code{id} or a \code{r3dmol} object (the output from
\code{r3dmol()})
}
\description{
Set the duration of the hover delay
}
