% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{m_set_preceived_distance}
\alias{m_set_preceived_distance}
\title{Set the distance between the model and the camera}
\usage{
m_set_preceived_distance(id, dist)
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})}

\item{dist}{Numeric distance.}
}
\value{
R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})
}
\description{
Essentially zooming. Useful while stereo rendering.
}
\examples{
library(r3dmol)

r3dmol() \%>\%
  m_add_model(data = pdb_6zsl, format = "pdb") \%>\%
  m_set_preceived_distance(dist = 200)
}
