% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{m_set_slab}
\alias{m_set_slab}
\title{Set slab of view}
\usage{
m_set_slab(id, near, far)
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})}

\item{near}{near clipping plane distance}

\item{far}{far clipping plane distance}
}
\value{
R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})
}
\description{
Set slab of view (contents outside of slab are clipped).
}
