% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{m_set_view_style}
\alias{m_set_view_style}
\title{Set global view styles}
\usage{
m_set_view_style(id, style)
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})}

\item{style}{css style object in list.}
}
\value{
R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})
}
\description{
Set global view styles
}
\examples{
library(r3dmol)

r3dmol() \%>\%
  m_add_model(data = pdb_6zsl, format = "pdb") \%>\%
  m_set_style(style = list(stick = list())) \%>\%
  m_zoom_to() \%>\%
  m_set_view_style(style = list(style = "outline", color = "blue"))
}
