% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{m_set_width}
\alias{m_set_width}
\title{Set viewer width}
\usage{
m_set_width(id, width)
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})}

\item{width}{Width \code{numeric} in pixels}
}
\value{
R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})
}
\description{
Set viewer width
}
