% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/example.R
\name{m_shiny_demo}
\alias{m_shiny_demo}
\title{Run examples of using \code{r3dmol} in a Shiny app}
\usage{
m_shiny_demo()
}
\description{
Run examples of using \code{r3dmol} in a Shiny app
}
\examples{
if (interactive()) {
  m_shiny_demo()
}
}
