% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{m_sphere_style}
\alias{m_sphere_style}
\title{Specify styling for Sphere}
\usage{
m_sphere_style(
  hidden = NULL,
  radius = NULL,
  scale = NULL,
  colorscheme = NULL,
  color = NULL,
  opacity = NULL
)
}
\arguments{
\item{hidden}{Boolean - do not show atom. Default \code{FALSE}}

\item{radius}{Override van der waals radius.}

\item{scale}{Scale radius by specified amount.}

\item{colorscheme}{Specify scheme to color the atoms by. Default is "default". Other choies are "Carbon", ssPyMOL", "ssJmol", "Jmol", "default", "amino",
"shapely", "nucleic", "chain", "chainHetatm", "prop".}

\item{color}{Discrete, fixed coloring, overrides any colorscheme.}

\item{opacity}{Opacity of spheres, 0 being invisible. Must be the same for
all atoms in the model.}
}
\value{
R3dmol \code{id} or a \code{r3dmol} object (the output from
\code{r3dmol()})
}
\description{
Returns a list for the \code{style = list(sphere = list())} function,
for setting the sphere specifications.
}
