% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{m_stick_style}
\alias{m_stick_style}
\title{Specify styling for Stick}
\usage{
m_stick_style(
  hidden = NULL,
  radius = NULL,
  singleBonds = NULL,
  colorscheme = NULL,
  color = NULL,
  opacity = NULL
)
}
\arguments{
\item{hidden}{do no show}

\item{radius}{radius of sticks}

\item{singleBonds}{Draw all bonds as single bonds if \code{TRUE}}

\item{colorscheme}{element based coloring}

\item{color}{fixed coloring, overrides colorscheme}

\item{opacity}{opacity, must be the same for all atoms in the model}
}
\description{
Returns a list for the \code{style = list(cartoon = list())} function,
for setting the stick specifications.
}
