% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/animation.R
\name{m_stop_animate}
\alias{m_stop_animate}
\title{Stop animation of all models in viewer}
\usage{
m_stop_animate(id)
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})}
}
\description{
Stop animation of all models in viewer
}
