% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{m_style_ballnstick}
\alias{m_style_ballnstick}
\title{Specifying Style for Ball and Stick}
\usage{
m_style_ballnstick(
  sphereScale = 0.3,
  colorScheme = NULL,
  color = NULL,
  sphereRadius = NULL,
  stickRadius = 0.25,
  hidden = FALSE,
  opacity = 1,
  colorfunc = NULL
)
}
\arguments{
\item{sphereScale}{Relative scaling of sphere atomic radii.}

\item{colorScheme}{Specify scheme to color the atoms by. Default is
"default". Other choices are "Carbon", "ssPyMOL", "ssJmol", "Jmol",
"default", "amino", "shapely", "nucleic", "chain", "chainHetatm", "prop".}

\item{color}{Specific coloring for everything in selection, overrides any
specified colorScheme.}

\item{sphereRadius}{Set absolute radius for all spheres in selection
(overriding the atomic raddii)l}

\item{stickRadius}{Set absolute radius for all sticks in the selection.}

\item{hidden}{Boolean - do not show atoms in selection. Default \code{FALSE}}

\item{opacity}{Opacity of spheres and sticks, 0 being invisible, 1 being
opaque. Must be the same for all atoms in the model.}

\item{colorfunc}{Allows the user to provide a function for setting the
colorSchemes, written in javascript.
\href{https://3dmol.csb.pitt.edu/doc/types.html#ColorschemeSpec}{Official
Documentation}}
}
\description{
Styling options for ball and stick representations.
}
\examples{
r3dmol() \%>\%
  m_add_model(data = pdb_1j72, format = "pdb") \%>\%
  m_set_style(style = m_style_ballnstick()) \%>\%
  m_zoom_to()
}
