% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{m_style_cartoon}
\alias{m_style_cartoon}
\title{Specify Styling for Cartoon}
\usage{
m_style_cartoon(
  color = NULL,
  style = NULL,
  ribbon = NULL,
  arrows = NULL,
  tubes = NULL,
  thickness = NULL,
  width = NULL,
  opacity = NULL
)
}
\arguments{
\item{color}{Strand color, may specify as 'spectrum' which will apply
reversed gradient based on residue number.}

\item{style}{style of cartoon rendering (trace, oval, rectangle (default),
parabola, edged)}

\item{ribbon}{whether to use constant strand width, disregarding secondary
structure; use thickness to adjust radius}

\item{arrows}{whether to add arrows showing beta-sheet directionality;
does not apply to trace or ribbon}

\item{tubes}{whether to display alpha helices as simple cylinders; does not
apply to trace}

\item{thickness}{cartoon strand thickness, default is 0.4}

\item{width}{cartoon strand width, default is secondary structure-dependent;
does not apply to trace or ribbon}

\item{opacity}{set opacity from 0-1; transparency is set per-chain with a
warning outputted in the event of ambiguity}
}
\description{
Returns a list for the \code{style = list(cartoon = list())} function,
for setting the cartoon specifications.
}
\examples{
r3dmol() \%>\%
  m_add_model(data = pdb_1j72, format = "pdb") \%>\%
  m_set_style(style = m_style_cartoon(color = "spectrum")) \%>\%
  m_zoom_to()
}
