% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{m_style_line}
\alias{m_style_line}
\title{Specify Styling for Lines}
\usage{
m_style_line(colorScheme = NULL, color = NULL, opacity = NULL, hidden = NULL)
}
\arguments{
\item{colorScheme}{Specify scheme to color the atoms by. Default is
"default". Other choies are "Carbon", ssPyMOL", "ssJmol", "Jmol",
"default", "amino",
"shapely", "nucleic", "chain", "chainHetatm", "prop".}

\item{color}{Fixed coloring, overrides \code{colorScheme}.}

\item{opacity}{Opacity, must be the same for all atoms in the model.}

\item{hidden}{Logical, do not show line.}
}
\description{
Returns a list for the \code{style = list(line = list())} function
for setting the line specifications.
}
\examples{
r3dmol() \%>\%
  m_add_model(data = pdb_1j72, format = "pdb") \%>\%
  m_set_style(style = m_style_line(color = "blue")) \%>\%
  m_zoom_to()
}
