% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{m_style_surface}
\alias{m_style_surface}
\title{Specify Styling for Surface}
\usage{
m_style_surface(opacity = NULL, colorScheme = NULL, color = NULL)
}
\arguments{
\item{opacity}{Opacity, 0 for transparent, 1 for opaque.}

\item{colorScheme}{Specify scheme to color the atoms by. Default is
"default". Other choies are "Carbon", ssPyMOL", "ssJmol", "Jmol",
"default", "amino",
"shapely", "nucleic", "chain", "chainHetatm", "prop".}

\item{color}{Fixed coloring, overrides colorScheme.}
}
\description{
Must be used inside \code{m_add_surface()} not any other styling function.
Returns a list for setting the surface specifications.
}
\examples{
r3dmol() \%>\%
  m_add_model(data = pdb_1j72, format = "pdb") \%>\%
  m_set_style(style = m_style_stick()) \%>\%
  m_add_surface(style = m_style_surface(opacity = 0.4)) \%>\%
  m_zoom_to()
}
