% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructor.R
\name{m_vector3}
\alias{m_vector3}
\title{Create a 3 dimensional vector}
\usage{
m_vector3(x = NULL, y = NULL, z = NULL)
}
\arguments{
\item{x}{x coordinate, \code{character} and \code{numeric} are both accepted.}

\item{y}{y coordinate, \code{character} and \code{numeric} are both accepted.}

\item{z}{z coordinate, \code{character} and \code{numeric} are both accepted.}
}
\value{
3 dimensional list object
}
\description{
Create a 3 dimensional vector
}
\examples{
library(r3dmol)
m_vector3(1, 2, 3)
}
