% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view_options.R
\name{m_viewer_spec}
\alias{m_viewer_spec}
\title{Specifying setup options for Viewer}
\usage{
m_viewer_spec(
  defaultcolors = NULL,
  nomouse = NULL,
  backgroundColor = NULL,
  hoverDuration = NULL,
  id = NULL,
  cartoonQuality = NULL,
  lowerZoomLimit = NULL,
  upperZoomLimit = NULL,
  antialias = NULL,
  orthographic = NULL,
  disableFog = NULL
)
}
\arguments{
\item{nomouse}{Whether to disable handling of mouse events. Disabled will
prevent user interaction.}

\item{backgroundColor}{color of the canvas's background.}

\item{id}{id of the canvas}

\item{cartoonQuality}{Defaults to 5}

\item{lowerZoomLimit}{Specify how far the user can zoom in.}

\item{upperZoomLimit}{Specify how far the user can zoom out.}

\item{antialias}{Logical whether to enable antialiasing, improving graphical quality. (can slow down large structures)}

\item{orthographic}{Logical. Setting orthographic instead of perspective representation.}

\item{disableFog}{Logical, disable fog, defaults to \code{FALSE}}

\item{defaultcolor}{Object defining default atom colors as atom => color property value pairs for all models within this viewer}
}
\description{
Returns a list for the setup \code{r3dmol()} function, to set
overall settings for the viewer going forward.
}
