% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pdb_1j72}
\alias{pdb_1j72}
\title{Crystal Structure of Mutant Macrophage Capping Protein (Cap G) with
Actin-severing Activity in the Ca2+-Free Form in PDB format}
\format{
PDB Format.
}
\source{
DOI: 10.2210/pdb1J72/pdb. \url{https://www.rcsb.org/structure/1J72}
}
\usage{
pdb_1j72
}
\description{
Crystal Structure of Mutant Macrophage Capping Protein (Cap G) with
Actin-severing Activity in the Ca2+-Free Form in PDB format
}
\keyword{datasets}
