% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pdb_6zsl}
\alias{pdb_6zsl}
\title{Crystal structure of the SARS-CoV-2 helicase at 1.94 Angstrom resolution in PDB format}
\format{
PDB Format.
}
\source{
DOI: 10.2210/pdb6ZSL/pdb. \url{https://www.rcsb.org/structure/6zsl}
}
\usage{
pdb_6zsl
}
\description{
Crystal structure of the SARS-CoV-2 helicase at 1.94 Angstrom resolution in PDB format
}
\keyword{datasets}
