% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r3dmol.R
\name{r3dmol-shiny}
\alias{r3dmol-shiny}
\alias{r3dmolOutput}
\alias{renderR3dmol}
\title{Shiny bindings for r3dmol}
\usage{
r3dmolOutput(outputId, width = "100\%", height = "400px")

renderR3dmol(expr, env = parent.frame(), quoted = FALSE)
}
\arguments{
\item{outputId}{output variable to read from}

\item{width, height}{Must be a valid CSS unit (like \code{'100\%'},
\code{'400px'}, \code{'auto'}) or a number, which will be coerced to a
string and have \code{'px'} appended.}

\item{expr}{An expression that generates a r3dmol}

\item{env}{The environment in which to evaluate \code{expr}.}

\item{quoted}{Is \code{expr} a quoted expression (with \code{quote()})? This
is useful if you want to save an expression in a variable.}
}
\description{
Output and render functions for using r3dmol within Shiny
applications and interactive Rmd documents.
}
