% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{removeAllModels}
\alias{removeAllModels}
\title{Delete all existing models}
\usage{
removeAllModels(id)
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})}
}
\value{
id R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})
}
\description{
Delete all existing models
}
