% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{removeAllSurfaces}
\alias{removeAllSurfaces}
\title{Remove all labels from viewer}
\usage{
removeAllSurfaces(id)
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})}
}
\value{
id R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})
}
\description{
Remove all labels from viewer
}
