% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/remove.R
\name{removeLabel}
\alias{removeLabel}
\title{Remove label from viewer}
\usage{
removeLabel(id, label)
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})}

\item{label}{R3dmol object label}
}
\value{
id R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})
}
\description{
Remove label from viewer
}
