% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{sdf_multiple}
\alias{sdf_multiple}
\title{Multiple sdf file example}
\format{
sdf format
}
\source{
\url{https://github.com/3dmol/3Dmol.js/blob/master/tests/test_structs/multiple.sdf}
}
\usage{
sdf_multiple
}
\description{
Multiple sdf file example
}
\keyword{datasets}
