% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setStyle.R
\name{setBackgroundColor}
\alias{setBackgroundColor}
\title{Set viewer background color}
\usage{
setBackgroundColor(id, hex, alpha)
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})}

\item{hex}{Hexcode specified background color, or standard color spec \code{character}}

\item{alpha}{Alpha level \code{numeric} (default \code{1.0})}
}
\value{
R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})
}
\description{
Set the background color (default white)
}
