% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{setZoomLimits}
\alias{setZoomLimits}
\title{Set lower and upper limit stops for zoom}
\usage{
setZoomLimits(id, lower = 0, upper = Inf)
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})}

\item{lower}{limit on zoom in (positive \code{numeric} number). Default \code{0}.}

\item{upper}{limit on zoom out (positive \code{numeric} number). Default \code{Inf}.}
}
\value{
R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})
}
\description{
Set lower and upper limit stops for zoom
}
