% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setStyle.R
\name{spin}
\alias{spin}
\title{Continuously rotate a scene around the specified axis}
\usage{
spin(id, axis = "y")
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})}

\item{axis}{Axis (\code{"x"}, \code{"y"}, \code{"z"}, \code{"vx"}, \code{"vy"}, \code{"vz"})
to rotate around. Default \code{"y"}. View relative (rather than model relative) axes are
prefixed with \code{"v"}.}
}
\value{
R3dmol id or a \code{r3dmol} object (the output from \code{r3dmol()})
}
\description{
Continuously rotate a scene around the specified axis
}
