% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setStyle.R
\name{stopAnimate}
\alias{stopAnimate}
\title{Stop animation of all models in viewer}
\usage{
stopAnimate(id)
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})}
}
\description{
Stop animation of all models in viewer
}
