% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{xyz_multiple}
\alias{xyz_multiple}
\title{Multiple xyz file example}
\format{
xyz format
}
\source{
\url{https://github.com/3dmol/3Dmol.js/blob/master/tests/test_structs/multiple2.xyz}
}
\usage{
xyz_multiple
}
\description{
Multiple xyz file example
}
\keyword{datasets}
