% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r3dmol.R
\name{init}
\alias{init}
\alias{r3dmol}
\title{Initialise a WebGL-based viewer}
\usage{
r3dmol(id = NULL, width = NULL, height = NULL, elementId = NULL)
}
\arguments{
\item{id}{HTML element id of viewer}
}
\description{
Create and initialize an appropriate viewer at supplied HTML element using specification in config
}
\examples{
library(r3dmol)
r3dmol() \%>\%
  m_add_model(data = pdb_6zsl, format = "pdb") \%>\%
  m_zoom_to()
}
