% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r3dmol.R
\name{init}
\alias{init}
\alias{r3dmol}
\title{Initialise}
\usage{
r3dmol(
  data_path = NULL,
  width = NULL,
  height = NULL,
  id = NULL,
  elementId = NULL
)
}
\arguments{
\item{data_path}{}
}
\description{
Initialise a 3d protein structure
}
