% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{m_add_arrow}
\alias{m_add_arrow}
\title{Create and add arrow shape}
\usage{
m_add_arrow(id, spec = list())
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})}

\item{spec}{Style specification}
}
\value{
R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})
}
\description{
Create and add arrow shape
}
\examples{
r3dmol() \%>\%
  m_add_model(data = pdb_6zsl, format = "pdb") \%>\%
  m_set_style(style = list(cartoon = list(color = 'spectrum'))) \%>\%
  m_add_arrow(
    spec = list(
      start = m_vector3(-10, 0, 0),
      end = m_vector3(0, -10, 0),
      radius = 1,
      radiusRadio = 1,
      mid = 1,
      clickable = TRUE,
      callback =
        "function(){
          this.color.setHex(0xFF0000FF);
          view.render()
        }"
    )
  )
}
