% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{m_add_box}
\alias{m_add_box}
\title{Create and add box shape}
\usage{
m_add_box(id, spec = list())
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})}

\item{spec}{Box shape style specification}
}
\value{
R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})
}
\description{
This method provides a shorthand way to create a box shape object
}
