% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{m_add_line}
\alias{m_add_line}
\title{Create and add line shape}
\usage{
m_add_line(id, spec = list())
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})}

\item{spec}{Style specification, can specify dashed, dashLength, and gapLength}
}
\value{
R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})
}
\description{
Create and add line shape
}
