% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{m_add_lines}
\alias{m_add_lines}
\title{Add Lines Between Points}
\usage{
m_add_lines(id, line_specs)
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from
\code{r3dmol()}).}

\item{line_specs}{a list of LineSpecs.}
}
\value{
R3dmol \code{id} or a \code{r3dmol} object (the output from
\code{r3dmol()})
}
\description{
Add lines between the given points.
}
\examples{
library(r3dmol)

line_specs <- list(
  list(
    start = m_sel(
      resi = 1:10,
      chain = "A"
    ),
    end = m_sel(
      resi = 1:10,
      chain = "B"
    ),
    dashed = TRUE
  ),
  list(
    start = m_sel(
      resi = 20:30,
      chain = "A"
    ),
    end = m_sel(
      resi = 20:30,
      chain = "B"
    )
  )
)

r3dmol() \%>\%
  m_add_model(data = pdb_6zsl) \%>\%
  m_set_style(style = m_style_cartoon()) \%>\%
  m_zoom_to() \%>\%
  m_add_style(
    sel = m_sel(resi = 1:10),
    style = c(
      m_style_stick(),
      m_style_sphere(scale = 0.3)
    )
  ) \%>\%
  m_add_lines(line_specs = line_specs)
}
