% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{m_add_res_labels}
\alias{m_add_res_labels}
\title{Add residue labels}
\usage{
m_add_res_labels(id, sel, style, byframe)
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})}

\item{sel}{Atom selection specification}

\item{style}{Style spec to add to specified atoms}

\item{byframe}{if true, create labels for every individual frame, not just current}
}
\value{
R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})
}
\description{
Add residue labels. This will generate one label per a
residue within the selected atoms. The label will be at the
centroid of the atoms and styled according to the passed style.
The label text will be \link[=resi]{resn}
}
\examples{
library(r3dmol)

r3dmol() \%>\%
  m_add_model(data = pdb_1j72, format = "pdb") \%>\%
  m_set_style(style = list(stick = list(radius = 0.15), cartoon = list())) \%>\%
  m_add_res_labels(
    sel = list(resn = "GLY"),
    style = list(
      font = "Arial",
      fontColor = "white",
      backgroundColor = "black",
      showBackground = TRUE
    )
  ) \%>\%
  m_zoom_to()
}
