% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{m_add_sphere}
\alias{m_add_sphere}
\title{Create and add sphere shape}
\usage{
m_add_sphere(id, spec = list())
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})}

\item{spec}{Sphere shape style specification}
}
\value{
R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})
}
\description{
This method provides a shorthand way to create a spherical shape object
}
\examples{
library(r3dmol)

r3dmol() \%>\%
  m_add_sphere(spec = list(
    center = m_vector3(0, 0, 0),
    radius = 10,
    color = "red"
  ))
}
