% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{m_add_unit_cell}
\alias{m_add_unit_cell}
\title{Create and add unit cell visualization}
\usage{
m_add_unit_cell(id, model, spec)
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})}

\item{model}{Model with unit cell information (e.g., pdb derived).
If omitted uses most recently added model.}

\item{spec}{Visualization style.}
}
\value{
R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})
}
\description{
Create and add unit cell visualization
}
\examples{
library(r3dmol)

r3dmol() \%>\%
  m_add_model(
    data = cif_254385,
    "cif",
    options = list(doAssembly = TRUE, normalizeAssembly = TRUE)
  ) \%>\%
  m_set_style(style = list(
    sphere = list(colorscheme = "Jmol", scale = 0.25),
    stick = list(colorscheme = "Jmol")
  )) \%>\%
  m_add_unit_cell(spec = list(
    alabel = "x",
    blabel = "y",
    clabel = "z",
    box = list(hidden = TRUE)
  )) \%>\%
  m_zoom_to()
}
