% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button.R
\name{m_button_add_res_labels}
\alias{m_button_add_res_labels}
\title{Add a button for adding residue labels}
\usage{
m_button_add_res_labels(
  id,
  sel = m_sel(),
  style = m_style_label(),
  label = "Show Labels",
  byframe = FALSE,
  hideButton = TRUE,
  hideLabel = "Hide Labels"
)
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from
\code{r3dmol()}).}

\item{sel}{Atom selection specification with \code{m_sel()}}

\item{style}{Style spec to add to specified atoms created with \code{m_style_label()}.}

\item{label}{String for button label.}

\item{byframe}{if TRUE, create labels for every individual frame in an
animation, not just the current.}

\item{hideButton}{Logical, whether to also create a button for removing all
labels from the viewer.}

\item{hideLabel}{String for hide labels button label.}
}
\value{
R3dmol \code{id} or a \code{r3dmol} object (the output from
\code{r3dmol()})
}
\description{
Add a button for adding residue labels
}
\examples{
# Add buttons to add and clear surface representations
r3dmol() \%>\%
  m_add_model(data = pdb_1j72, format = "pdb") \%>\%
  m_zoom_to() \%>\%
  m_set_style(m_style_cartoon()) \%>\%
  m_button_add_res_labels(sel = m_sel(resi = c(70, 80, 90)))
}
