% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button.R
\name{m_button_add_style}
\alias{m_button_add_style}
\title{Add button for Adding a Style}
\usage{
m_button_add_style(
  id,
  style = m_style_cartoon(),
  sel = m_sel(),
  label = "Style"
)
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from
\code{r3dmol()}).}

\item{style}{Style spec to apply to specified atoms using \code{m_style_*()}.
Multiple styles can be supplied inside a vector (i.e.)
\code{c(m_style_cartoon(), m_style_stick())}.}

\item{sel}{Atom selection specification with \code{m_sel()}}

\item{label}{String for button label.}
}
\value{

}
\description{
Creates a button that will add the given style to the selection when pressed.
}
