% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/button.R
\name{m_button_spin}
\alias{m_button_spin}
\title{Add a button for spinning the scene}
\usage{
m_button_spin(
  id,
  speed = 1,
  axis = "y",
  label = "Spin",
  stopButton = TRUE,
  stopLabel = "Stop"
)
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from
\code{r3dmol()}).}

\item{speed}{Speed multiplier for spin animation. Defaults to 1. Negative
value reverses the direction of spin.}

\item{axis}{Axis (\code{"x"}, \code{"y"}, \code{"z"}, \code{"vx"},
\code{"vy"}, \code{"vz"}) to rotate around. Default \code{"y"}. View
relative (rather than model relative) axes are prefixed with \code{"v"}.}

\item{label}{String for button label.}

\item{stopButton}{Logical, whether to also create a button to stop the spin.}

\item{stopLabel}{String for the stop button label.}
}
\value{

}
\description{
Add a button for spinning the scene
}
