% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{m_render}
\alias{m_render}
\title{Render current state of viewer}
\usage{
m_render(id)
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})}
}
\description{
Render current state of viewer, after adding/removing models,
applying styles, etc. In most cases, the model will render automatically,
only call it when manual rendering is required.
}
\examples{
library(r3dmol)

r3dmol() \%>\%
  m_add_model(data = pdb_6zsl, format = "pdb") \%>\%
  m_render()
}
