% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/set.R
\name{m_set_style}
\alias{m_set_style}
\title{Set style properties to all selected atoms}
\usage{
m_set_style(id, sel = list(), style = list())
}
\arguments{
\item{id}{R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})}

\item{sel}{Atom selection specification}

\item{style}{Style spec to apply to specified atoms}
}
\value{
R3dmol \code{id} or a \code{r3dmol} object (the output from \code{r3dmol()})
}
\description{
Set style properties to all selected atoms
}
\examples{
library(r3dmol)
r3dmol(data_path = pdb_6zsl) \%>\%
  m_set_style(style = list(cartoon = list())) \%>\%
  m_set_style(sel = list(chain = "A"),
           style = list(stick = list(
             radius = 0.5,
             colorscheme = "magentaCarbon"
           )))
}
