% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{m_shape_spec}
\alias{m_shape_spec}
\title{Specify Styling for Generic Shapes}
\usage{
m_shape_spec(
  color = NULL,
  alpha = NULL,
  wireframe = NULL,
  hidden = NULL,
  frame = NULL,
  clickable = NULL,
  callback = NULL
)
}
\arguments{
\item{color}{Solid color values.}

\item{alpha}{Transparency value. 1 for opaque, 0 for invisible.}

\item{wireframe}{Draw as wireframe, not solid surface.}

\item{hidden}{If true, do not display object.}

\item{frame}{If set, only display in this frame of an animation.}

\item{clickable}{If true, user can click on object to trigger callback.}

\item{callback}{Function to call on click.}
}
\description{
Specify Styling for Generic Shapes
}
\examples{
r3dmol() \%>\%
  m_add_model(data = m_fetch_pdb("1bna")) \%>\%
  m_add_sphere(
    center = m_sel(resi = 1),
    spec = m_shape_spec(color = "green", wireframe = TRUE)
  ) \%>\%
  m_zoom_to(sel = m_sel(resi = 1))
}
