% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{m_style_sphere}
\alias{m_style_sphere}
\title{Specify Styling for Sphere}
\usage{
m_style_sphere(
  scale = 1,
  colorScheme = NULL,
  color = NULL,
  radius = NULL,
  hidden = FALSE,
  opacity = 1
)
}
\arguments{
\item{scale}{Scale radius by specified amount.}

\item{colorScheme}{Specify scheme to color the atoms by. Default is
"default". Other choices are "Carbon", "ssPyMOL", "ssJmol", "Jmol",
"default", "amino",
"shapely", "nucleic", "chain", "chainHetatm", "prop".}

\item{color}{Discrete, fixed coloring, overrides any colorScheme.}

\item{radius}{Override van der waals radius.}

\item{hidden}{Boolean - do not show atom. Default \code{FALSE}.}

\item{opacity}{Opacity of spheres, 0 being invisible. Must be the same for
all atoms in the model.}
}
\description{
Styling options for the sphere representation. Used inside
\code{m_add_style()} and \code{m_set_style()}.
}
\examples{
r3dmol() \%>\%
  m_add_model(data = pdb_1j72, format = "pdb") \%>\%
  m_set_style(style = m_style_sphere(radius = 0.5)) \%>\%
  m_zoom_to()
}
