% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style.R
\name{m_style_sphere}
\alias{m_style_sphere}
\title{Specify Styling for Sphere}
\usage{
m_style_sphere(
  hidden = NULL,
  radius = NULL,
  scale = NULL,
  colorscheme = NULL,
  color = NULL,
  opacity = NULL
)
}
\arguments{
\item{hidden}{Boolean - do not show atom. Default \code{FALSE}.}

\item{radius}{Override van der waals radius.}

\item{scale}{Scale radius by specified amount.}

\item{colorscheme}{Specify scheme to color the atoms by. Default is "default". Other choies are "Carbon", ssPyMOL", "ssJmol", "Jmol", "default", "amino",
"shapely", "nucleic", "chain", "chainHetatm", "prop".}

\item{color}{Discrete, fixed coloring, overrides any colorscheme.}

\item{opacity}{Opacity of spheres, 0 being invisible. Must be the same for
all atoms in the model.}
}
\value{
List of values for input inside \code{style = list()} for sphere.
}
\description{
Returns a list for the \code{style = list(sphere = list())} function,
for setting the sphere specifications.
}
\examples{
r3dmol() \%>\%
  m_add_model(data = pdb_1j72, format = "pdb") \%>\%
  m_set_style(style = m_style_sphere(radius = 0.5)) \%>\%
  m_zoom_to()
}
