% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/view-options.R
\name{m_viewer_spec}
\alias{m_viewer_spec}
\title{Specifying setup options for viewer}
\usage{
m_viewer_spec(
  id = NULL,
  defaultcolors = NULL,
  cartoonQuality = 5,
  antialias = TRUE,
  nomouse = FALSE,
  backgroundColor = "white",
  lowerZoomLimit = 5,
  upperZoomLimit = 400,
  orthographic = FALSE,
  disableFog = FALSE
)
}
\arguments{
\item{id}{id of the canvas.}

\item{defaultcolors}{Object defining default atom colors as atom => color
property value pairs for all models within this viewer.}

\item{cartoonQuality}{Defaults to \code{5}.}

\item{antialias}{Logical, disable to decrease quality but improve
performance.}

\item{nomouse}{Whether to disable handling of mouse events. Disabled will
prevent user interaction.}

\item{backgroundColor}{color of the canvas's background.}

\item{lowerZoomLimit}{Specify how far the user can zoom in.}

\item{upperZoomLimit}{Specify how far the user can zoom out.}

\item{orthographic}{Logical. Setting orthographic instead of perspective
representation.}

\item{disableFog}{Logical, disable fog, defaults to \code{FALSE}}
}
\description{
Returns a list for the setup \code{r3dmol()} function, to set
overall settings for the viewer going forward.
}
